+++ src/components/Web3Provider/index.tsx
@@ -2,8 +2,8 @@ import { Web3ReactHooks, Web3ReactProvider } from '@web3-react/core'
 import { Connector } from '@web3-react/types'
 import { Connection } from 'connection'
 import { getConnectionName } from 'connection/utils'
-import useOrderedConnections from 'hooks/useConnectors'
 import useEagerlyConnect from 'hooks/useEagerlyConnect'
+import useOrderedConnections from 'hooks/useOrderedConnections'
 import { ReactNode } from 'react'
 
 export default function Web3Provider({ children }: { children: ReactNode }) {
